Try to be smarter about dllPath handling.
If dllPath does not point to a real directory (i.e. it's the default and we're
on Windows), then leave the DLLs with relative paths and let IKVM try to figure
out where they come from. This will hopefully allow the standard libs to be
properly resolved without having to hardcode the path to your standard C#
directories if you're using this on Windows.